[alloc @ alla@mint-20060514171948-be943ed1763aac05] - Readjusted size of menu, made it use table_box - Pruned toolbar settings from stylesheets - Added pending amount to tf box on home page - Added check that user has correct perms for deleting a tf - Fixed invoicesUpload, invoiceItem, searchInvoices, exp-one-off for page globals - Fixed transaction, searchTransactions, reconciliationReport, wagesUpload for page globals - Made it so that regular users can do invoice and transaction searches without running into   permission hiccups, by restricting their searches to their tfs - Killed email sending for item loans, fixed itemLoan, item page globals

+++ finance/lib/tf.inc.php
@@ -67,14 +67,19 @@ class tf extends db_entity {
 
 
   function delete() {
+    global $current_user, $TPL;
     $db = new db_alloc;
 
-    $query = "DELETE FROM transaction WHERE tfID='".$this->get_id()."'";
-    $db->query($query);
+    if ($current_user->have_role("god") || $current_user->have_role("admin")) {
+      $query = "DELETE FROM transaction WHERE tfID='".$this->get_id()."'";
+      $db->query($query);
 
-    $query = "DELETE FROM tfPerson WHERE tfID='".$this->get_id()."'";
-    $db->query($query);
-    db_entity::delete();
+      $query = "DELETE FROM tfPerson WHERE tfID='".$this->get_id()."'";
+      $db->query($query);
+      db_entity::delete();
+    } else {
+      $TPL["message"] = "Permission denied.";
+    }
   }
